+++ src/models/MetaModel/index.ts
@@ -1,1 +1,3 @@
 export { default as MetaModel } from './MetaModel';
+
+export * from './types';