+++ src/types/ILocalStore.ts
@@ -1,0 +1,4 @@
+export interface ILocalStore {
+  // Любой локальный store должен реализовывать метод destroy, в котором происходят отписки от реакций
+  destroy: VoidFunction;
+}