+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v2.2.0
+
+- [+] модели для работы со списками: ApiListModel и MockApiListModel
+- [+] PollingModel
+- [+] BlurAndFocusHandlerModel
+- [+] модели для таймеров: BaseTimerModel, TimerModel, FixedTimerModel
+
 ## v2.1.0
 
 - [*] Sentry version